tmf: provide API to add tracing nature to other project types
This creates a nested shadow tracing project with the name
.tracecompass-<project-name>. The shadow project is create under the
hidden directory .tracecompass in the project root. This will be as
to show the TmfProjectElement as root node in the Project Explorer of
this project. The name "Trace Compass" and the Trace Compass icon will
be displayed. The Experiment and Traces folder are visible underneath.

All trace and experiment resources are created and handled in the
shadow project, but the TMF project model elements are shown in the
current project. The shadow project is hidden from the user by default.

Change-Id: I376a87c45b195ffa4d3416266bf5709530809014
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/88789
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>