datastore: Add a Predicate for children node header
This adds the possibility to filter children nodes by their header data
from a Predicate on the intervals. There is a new method in the tree and
node to transform a Predicate on the interval to a Predicate on the core
node data. Each tree and node implementations are responsible for their
own data.

Change-Id: Ibe44b089a8d25e3f4b53718570372c96e150c1ed
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/89120
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-by: Hudson CI