releng: Add missing javax.xml dependency for staging target
This fixes the missing javax.xml dependency for org.apache.xerces.
javax.xml is no longer packaged with the Eclipse RCP features but
apache.xerces still depend on it (for XML editor, etc).
If Orbit updates the dependency then we'll be able to remove it but
it might not be worth the effort at this time.

See also
https://bugs.eclipse.org/bugs/show_bug.cgi?id=467028

Change-Id: I9c1fa21efb5fa992464cae2af08641c80dde23f3
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/90886
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>