releng: Use Eclipse 4.6 milestone builds
This will allow us to test and build using the Eclipse 4.6.3 RCs.
Also, derby can now be used from Orbit because test.performance was
updated to use “import package” instead of require-plugin. It is better
to use the one in Orbit because it is on download servers instead of
build.eclipse.org which has proven to be unreliable.

Change-Id: Ie61da4b0cec68e5071ef529993e8d9815a85b523
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/91153
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>