datastore: Add API for single queries
When what we want is only one interval, there is no need to get them
all, we can just fast return when the one interval is found.

This change alones reduces the performance impact of the datastore by
~2x for single queries in the case of smallish traces.

This patch also adds a binary search to find the first interval to
look at.

Change-Id: I77d0ac9818f272f7d160f92666445ce2d8b3200a
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/91753
Reviewed-by: Hudson CI
Reviewed-by: Loic Prieur-Drevon <loic.prieur.drevon@ericsson.com>