datastore: Use time conditions with long
Using primitive type long instead of a generic Long for range conditions
reduces the need for boxing and unboxing time values.

This change alone reduces the performance impact of the datastore by ~2
in use cases of small size for both single and full queries. Coupled with
the patch with single queries API, it reduces the performance impact of
single queries to < 100%

Change-Id: I63ebb20b2441aff1345b733f1ffc25314e63ffd2
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/91754
Reviewed-by: Hudson CI
Reviewed-by: Loic Prieur-Drevon <loic.prieur.drevon@ericsson.com>