xml: Move xml strings and schemas to public packages
Since the XML is now extendable, external plugins may need to make use
of those values, so they are now part of the API. As for the schema,
though not mandatory to run correctly, it makes it obvious that they
can be used by external plugins.

Also, a TmfXmlUtils class has been added to the API and the
getChildElements method from XmlUtils has been moved there.

Change-Id: If1a98badcf52cbbafe77699d0cf6c6e9df3b47c8
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/92276
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Hudson CI