tmf: Bug 506095: Fix NullPointerException in XML views
When the XML Time Graph View or XML XY Chart View is opened from the
Window > Show View > Other... menu, it does not have a secondaryId. This
should not cause an exception, and the view should remain blank.

Additionally, the view should then not wait for the viewInfo name to be
set, it will never happen and the build job or line chart update thread
remains stuck forever. As a consequence, the viewInfo should be
initialized before creating the view controls.

Change-Id: I0d21cbcc02505bd5d83b7fb4335bf780943fe136
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/92736
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI