tmf: Bug 506660: Fix zoom out from 1 ns window range in time graph
Make sure that zooming out from 1 ns window range is possible with the
Zoom Out tool bar button.

Fix double to long rounding error when zooming out.

Make the 'focused' time (selection if visible, center of previous window
range otherwise) centered in the new window range when zooming out from
1 ns window range.

Keep the selection as 'focused' time when it is exactly one of the
previous window range bounds.

Snap the window range to the time bounds when zooming out of range.

Change-Id: I4501d178acf1c1383caecc95d5bf2b3ec534e602
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/93066
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>