timing.core/ui: make flamegraphs aggregate threads by default
If you have a flow of data with several threads using a scatter/
gather pattern, the flamegraph aggregates it to show the global
picture of all the threads.

To switch from the per-thread (default) to aggregate (classic) mode
and back, click on the drop-down menu and go to "Content Presentation".

Change-Id: I1fab0e09278d029302129071af31b1d809810b0a
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/93589
Reviewed-by: Hudson CI