tmf: Fix time graph scale ticks and grid lines
The time graph control depends on the time graph scale to get the tick
list of x-coordinates to use them to draw vertical grid lines. The tick
list is computed when the time graph scale is painted.

It appears that the order in which these two controls are painted is not
guaranteed. Therefore time graph control can sometimes paint grid lines
using the previous time range's tick list. With AbstractTimeGraphView
this usually gets fixed after a short delay when the zoom thread
triggers another repaint at the same window range, but not always.

This tick list in the time graph scale is no longer cached, and is
always recalculated from the current window range when requested by the
time graph control.

Another issue is that the last grid line disappears when its
corresponding tick label does not have room to draw completely. It is
now always drawn, and the tick label is simply truncated.

The calculation of the number of tick label digits for the CALENDAR time
format is fixed. It was too large and inconsistent, which could cause
grid lines to be unnecessarily far apart.

The digit width is cached so that it can be used outside of the time
graph scale's paint event.

Change-Id: Iaaa1caae8328d469700a7ad8aee726795b356930
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/93620
Reviewed-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>