tmf: Fix abstract time graph view end bound
Some abstract time graph view implementations set the end time bound of
the view to the state system end time + 1. This was to be able to see
the last state in the case that the last event in the trace resulted in
a state change (the time graph event begins at end time and has a
duration of 1).

However the TmfTraceManager does not allow the window range to exceed
the trace bounds. So when the view window range was changed to show the
view end bound, it immediately snapped back to the trace end time. This
can be seen by double-clicking on the time graph scale to show the full
range, or by zooming in and scrolling all the way to the end.

This effect also prevents the view from zooming out of a 1 ns window
range at the trace end time.

The end time bound can be set to the state system end time directly. The
final state can still be seen, selected and hovered for its tool tip,
although it has a 1 pixel width.

Change-Id: I6e65fc2f960c54af0c53e4725dba334ba30fe827
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/93649
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>