releng: add script for creating simRel build
This script is currently used on the eclipse server to create simRel
milestone builds and simRel releases. It needs to be run on the
remote target. Using a Hudson job will make it easier to run it
with the correct command-line parameters.

With this commit it's integrated into version control.

Change-Id: Ida9431833ff0109843a13ee63b7f2bcf93cdf16e
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/93831
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>