tmf: Fix BufferedRandomAccessFile
The method read(byte[]) was not overloaded and therefore reading
natively at the real file pointer, which could be different than the
buffered file pointer set at the last seek or read. The method now reads
from the buffer.

The write methods were not overloaded, and had the same issues. The real
file pointer is now synched with the buffered file pointer before the
write, and the buffered file pointer is synched with the real file
pointer after the write.

When repeatedly seeking backwards and reading, the buffer was
invalidated and filled at every iteration. The file will now position
itself to read ahead by half of the buffer size when seeking backwards.
This will improve the chance to benefit from the read buffer.

Unit tests are added.

Change-Id: If9662844818f1500f8d403b1120083892e788188
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/93840
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>