common.core: Add monitoring to logger
Scope logs and flow scopes now publish results to JMX listeners.

In order to test this, you can connect to Trace Compass with jconsole or
Visualvm and look at the mbeans, a TraceCompassMonitor bean should be
there. As more tracing is added, this will be more populated. It is
interesting to monitor like this if we ever decide to go for a
client-server architecture.

By debugging, it appears the code is unused when the beans are not
connected, therefore there should be no (observable) slowdown to this
code. For preciseness, upon connecting the beans, the
TraceCompassMonitor getters are only then called. The rest of the hereby
changed code is however hit upon running either the application, or
Incubator's trace-server while running [1]'s tests.

[Added] Publish logs to JMX to monitor performance

[1] https://github.com/eclipse-cdt-cloud/tsp-python-client

Change-Id: I2966eec7c43e29b4498b82aab72e26dccc9ddc1a
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/94018
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>