tmf: Fix experiment trace time range in Project Explorer
When the supplementary files are deleted, the start and end times stored
in the corresponding TmfTraceElement are reset. However this does not
affect the trace element instances of the same trace that can be present
under an experiment element. These instances keep their own start and
end times that could be obsolete.

The navigator label provider should simply always use the corresponding
trace element under the traces folder for trace elements under
experiments.

Change-Id: Ia7f13aab77d7b565a374a339a6c2e9b2c6347978
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/94604
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>