tmf: Support nested experiments in core
This patch adds support for experiments within experiments in the core
plug-in. It is not yet possible to create nested experiments with the
Project Explorer UI, but this patch adds support for nested experiments
that can be created by plug-in extensions programmatically.

The method TmfTraceManager.getTraceSet() is now recursive and returns
only the leaf traces of a given experiment.

The method TmfTraceManager.getTraceSetWithExperiment() is now recursive
and returns all child experiments and traces of a given experiment,
including itself.

The method TmfTrace.traceOpened() now executes (e.g. analyses) when the
trace being opened is any parent of the trace instance.

Change-Id: I2afd97cd482e6724dfaf2d0eb687cbe81de2790c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/94773
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>