symbol provider: Find symbols not at direct address
In some use cases like sampling, the symbol most often does not hit a
value that is mapped. In those cases, the symbol should still be
resolved to the floor value of the map. This is possible since the
symbol map is complete for a given executable. The sampling will work as
long as it is in a mapped region.

There is also a method to specify a suffix for symbols marking the end
of a region, to avoid symbols out of the region to resolve to the last
symbol of the previous region.

Change-Id: I97de1899bc1fd4488a356257fc61a79d51f1c29e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/94778
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>