tmf: Split ISymbolProvider into core and UI parts
The ISymbolProvider interface and related classes were all living in the UI
plugin only because of the preference page option, which is a UI-only
construct. This limitation was known since the inception of the symbol
provider, with the intent of eventually correctly splitting it into
core and UI components. This patch partially fullfills that.

Future work:

* Make symbol providers in core.

Change-Id: I64696929fcd8f47ba34b6a7b48d5e814faca7616
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/95115
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI