tmf.core/ui: Move some logic of the symbol provider to core
This moves the Symbol provider to core. Special care must be taken to
ensure that the preference page is available if required.

Moves callstack symbol providers to core interfaces.

Makes flamegraph use symbol aspect.

Most changes are internal, so this does not affect the API much.
Also, now CallgraphAnalysis can be in Core.

Change-Id: Ie4c37b670f599fad8fb74e08b458c0db65bb312c
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/95118
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>