XML: Add support for text in time graph events
This adds an attribute to the entry in the view's XML description. This
attribute, if true, will cause the text for every time event to be
displayed. It is all or nothing, per entry basis, but it is simple to
use and answers many use cases.

This does not exclude eventually adding more complex behavior for text
(the control flow view's system call text is an example).

Change-Id: I95945d3595723fc70408eeddf8fa81073a527394
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/95380
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Hudson CI