timing: Listener on table viewer should not be mandatory
Race conditions can happen in views where there is a segment store table
and another viewer (like density view) where the data is updated twice
in the table, one by the table viewer itself and one by the density
viewer.

Now the density views can specify not to add a listener for the table
and let the density viewer update the data.

Change-Id: I1d12560bf6dd5f43a667be5836176b4af6f897bc
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/95406
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI