tmf.ui: Support secondary ID for TmfAnalysisViewOutput
This adds the possibility to add the analysis ID as secondary ID when
constructing the TmfAnalysisViewOutput objects.

This is useful for views that support a type of analysis instead of
only one analysis. Analyses of this type will associate the view and
the analysis ID to obtain a view output that will create a different
view for each analysis.

For example, each callstack analysis under a trace or experiment could
automatically have their own flamegraph associated.

Change-Id: Ife0da12d20ab7f77649d00fb580997ff3acd80a6
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/95832
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>