releng: Fix ProjectExplorerTracesFolderTest
The generated clashing LTTng traces in TestDirectoryStructureUtil are
changed to have a duration of 1001 ns (not 1000 ns) for 1001 events.

This prevents the time delta between events from flooring down to 0 ns,
which causes all events and all packet begin and end times to be equal.

There is probably another bug in the CTF reader that seeks to the
incorrect packet when there are multiple packets with the same begin and
end times.

Change-Id: Icea44a26f9b4913c2bee7d0bddae575a7263b58f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/95961
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>