tmf.ui: Fix the destination group in the import wizard
The destination group was handle by the parent class and add a browse
button and the possibility to write the destination path. The text field
and the browse button does nothing in the current version of Trace
Compass and it is a bit confusing since these part are not deactivate. I
removed the browse button and set the text field to read-only.

Change-Id: I11f0fe66609792785c44e9af19868ca1f5cccb82
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/96306
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>