ss: Use Objects in StateSystem instead of ITmfStateValue
Previously, the state system would store intervals' values as
ITmfStateValues, which would just encapsulate {null, int, long,
double, string} types.
This patch replaces ITmfStateValues by Objects for the values,
so that the representations have one less object to create/GC.

This does not break the API as the intervals can still return
ITmfStateValues, which are lazily created.
Makes StateSystem 10-25% faster.
Makes it easier to Implement new payloads in the ss.

Change-Id: I4065ae8163ae405b4e0b9ff8bc402502aee44d9b
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/96471
Reviewed-by: Hudson CI