tmf: Add Open As Experiment... item to Project Explorer context menu
The user can select one or many traces and/or trace folders to create
and open an experiment of the selected experiment type. All selected
traces and traces recursively found in selected trace folders are added
to the experiment.

If an experiment with the same name and traces already exists, it is
reopened (or selected if it is already opened). Otherwise, a new
experiment is created and opened.

Change-Id: Ic45e55fc0ad0d7830a803811c486ebab29fc4467
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/96688
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Hudson CI