ust: Add trace constructor with event factory
Traces extending the UST trace may pass their own event factory to the
constructor to parse additional fields.

Example use case for this: A JUL trace of Trace Compass taken with LTTng
is a UST trace, but the CTF field 'msg' is a json string from which we
can extract extra fields. The event factory will do just that.

Change-Id: Ic9ac1fde9df655f6b529cb0fbcd2ed0d3932d0b5
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/96765
Reviewed-by: Hudson CI