xml.ui: Add label prefix for XML Latency Views
This patch adds an element to the <header> of the xml analysis
description The new element is a prefix for the XML analyses latency
views. The new name of the latency table will be like "PREFIX_LABEL
Latency Table".
This could be easily extendable to all type of xml view.

Change-Id: I697424b2e5d9ac4a15d96760fd0ca5011159a93e
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/96966
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>