ust: Add a LinuxTidAspect using the context._vtid
And have the callstack analysis use the aspect instead. This will allow
the callstack analysis to work when an aspect discovering the TID from a
kernel trace is added (in the incubator).

The callstack analysis requirement will need to be updated for it to
work without the context vtid field.

Change-Id: I8b31acfd0c6b03685403c57c650fdb3474e50c26
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/98250
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>