ctf: Allow a TSDL symbol to be used in several places.
This is a problem highlighted with LTTng 2.9. A TSDL symbol
(e.g. a struct or something) can be used in several places
at once, so the symbol should not be re-used as the parent is
re-assigned, rather, the parent should be immutable and the
child should be duplicated if added elsewhere.

This also does not allow a scope to contain its parent.

Fixes bug: 517830

Change-Id: I8f818a3cff33c4bca8d7d22203b726df6b9659a2
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/98645
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI