tmf: Bug 518081: Trace elements without trace type should not have child
Make sure that default children of TmfTraceElement are added or removed
based on the presence of a valid trace type in refreshChildren().

Show the trace 'type ID' in Properties view even if it not a valid one.
In that case the 'type' name is left blank.

Cleanup the new or modified trace types when custom parser wizard is
finished or when custom parsers are imported, to refresh existing traces
that gain a valid trace type.

Change-Id: I9ff0b3af17c0674c9d9c0700805b394e83e63aec
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/99042
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>