tmf: Improve import operation monitor task name handling
When iterating through file elements to determine their trace type,
change the task name to 'Examining <path>'.

When iterating through file elements to determine if they are an archive
that should be extracted, change the task name to 'Examining <path>'.

This prevents the task name from lingering as 'Importing <path>' for the
last imported file while other files are being processed.

Subtasks are suppressed when importing file elements since the task name
already provides the imported path.

The subtask is cleared after extracting archives so that the last
extract file subtask name does not linger for subsequent operations.

Change-Id: I311c1f8e106df1928c387296da9029aa1e4a55cf
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/99148
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>