graph: Pass graph module to critical path constructor
This avoid a parameter change event when the trace is set and thus does
not schedule the critical path at that point. Besides, in the case of
the critical path module being a sub-module of the graph module, it
makes sense that it should know about its parent module at construction
time.

Maybe if the graph is not set anywhere anymore we can remove the graph
parameter altogether and just leave it in the constructor?

Bug: 518264

Change-Id: Ifb684df234c8e2fe8f38661fdcca043dab151fec
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/99337
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>