tmf.core : Create models for XY and introduce data providers
This is a major change in the architecture for the XY viewers.
With this start of work, UI and Logic will be decoupled.

Viewers only know a DataProvider that returns a Model. The providers
know relevant analysis and will build the model. This architecture
will be, in the future, generalized other viewers.

Change-Id: I52414f5aa48648b7e7f99e3e0196af115877e6ac
Signed-off-by: Yonni Chen <yonni.chen.kuang.piao@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/100487
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>