tmf.xml: Add new state value type 'script'
A script state value can have many child state values underneath, their values
can be used in the script by specifying a unique id for each one and using
them as variables in the script.

At this moment the script should be passed directly in the value of the script
state value element.

Only JavaScript is supported at the moment.

Change-Id: Ie92f20c6910d9f730585dfd8cbc1017bd5b0cb57
Signed-off-by: Abderrahmane Berhil <abdelrahmane.b@gmail.com>
Reviewed-on: https://git.eclipse.org/r/99754
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>