ctf.core: Read traces with other files in the ctf directory
Before the trace would fail to open noting: "Magic mismatch".
This patch simply does not add the file to the readers if there
is a magic mismatch.

One test has changed because of this. Adding an invalid stream no
longer causes an exception to be thrown, rather a file is simply
ignored.

This is a bit of a grey zone in terms of the CTF contract as the
spec states that a CTF directory has one or more event stream and
one metadata file. It does not state that this is the only thing
the trace contains.

Change-Id: If9269c66216d1435f86dec410e5b039e60f00f0e
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/100946
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-by: Hudson CI
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>