ss: Add quark dimension to narrow down search queries and rsht
The State History Tree was considered to only have the time
dimension. However most queries ask for specific quarks too.
We can benefit from using the range of quarks to further
narrow down the nodes which need to be searched in the sht.

This patch adds the quark bounds to node headers and adapts
relevant query methods.

This makes single queries and 2D queries up to 50% faster.

Change-Id: Id029e4fe56e946f40f1d05a25b6c398e00863007
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/70903
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>