tmf: Prevent unnecessary modification of index files
The abstract file checkpoint collection now manages a dirty flag. Only
when the collection is modified, the file header is written as invalid
and the dirty flag is set. When disposing the collection, the valid file
header is serialized only if the dirty flag is set.

When re-opening and disposing a valid file for which no change is
detected by the indexer, the file will no longer be modified.

Change-Id: I9e7a268376b7e28c05b2caa78c3aa3077054178e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/101229
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-by: Hudson CI