tmf: Improve project model refresh when trace opened
The project registry instance is now the single object that needs to
handle the trace opened signal in order to refresh the necessary model
elements (e.g. to update the analysis elements under the trace).

This avoids the need for every trace and experiment element to register
with the signal manager and to handle the trace opened signal.

The parent of the opened trace element is no longer refreshed.

The method TmfProjectRegistry.findElement(IResource) is added.

Change-Id: I7641f6e337dc37904e4a7757cb37cae42ea930e6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/101283
Reviewed-by: Hudson CI
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>