tmf: Refresh only affected model elements
When a resource change event is detected, the project model is refreshed
only for the affected elements, i.e. the model element if the resource
content has changed, or its parent if the resource is added or removed.

The common viewer is refreshed for the whole affected project, since a
resource change (e.g. supplementary files) could affect the
visualization of many elements.

When the trace element bounds are modified, only the viewer needs to be
refreshed.

Change-Id: Id6a741b1cba2af31880db568c9e58365e9f77eea
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/101368
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Hudson CI