tmf: Update refresh handler activation
The RefreshHandler and its command are now visible and enabled when the
selection contains any type of project model element, and multiple
selection is now supported.

The RefreshTraceContentHandler and its command are now visible and
enabled only when the active part is a trace editor.

As a consequence, invoking the Refresh command (F5) when a trace element
is selected in the Project Explorer view now executes the RefreshHandler
instead of the RefreshTraceContentHandler.

Change-Id: I0bfd424647a941b941115f19af6e450f9e4dfe5b
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/101561
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-by: Hudson CI