Tmf.ui: Fix FindTest issues (fDirty and exceptions)
If dirty goes below 0, log an error, the dirty mechanism checks if
dirty == 0, if decrement is called but increment was not, it will
only be not dirty when it is actually dirty. This can happen if there
is an exception thrown in the code. Decrements are now placed in
finally blocks.

Also ControlflowviewFindTest and ResourceViewFindTest time out while
waiting for the find shell. It appears this is due to the trace taking
too long to load. A shorter trace is now used for the test.

Side-bonus: the build time should be a bit faster.

Change-Id: I50efe6412c1bbfc76426722b1718703b68ac1140
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/101770
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI