tmf: Add properties sub-folder to supplementary folder
The sub-folder '.properties' is added to every trace's supplementary
folder. The folder is set as hidden, so any files stored in this folder
do not get deleted by method deleteSupplementaryResources() when a trace
is modified by an action that normally requires the supplementary files
to be rebuilt. The hidden folder is not visible in the Delete
Supplementary Files dialog.

Change-Id: Idd91a9b6d198dcc6118f672a308849fcf372e54d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/102027
Reviewed-by: Hudson CI