tmf: Propagate deletion of supplementary resources to experiments
The method deleteSupplementaryResources() is called when a trace
is modified by an action that normally requires the supplementary files
to be rebuilt.

It should be assumed that any experiment that contains this trace also
needs its supplementary files to be rebuilt. The deletion of
supplementary resources is therefore propagated to those experiments.

Change-Id: I16ab5818d7bf1c9989fd86a119b6122cd4ea9d74
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/102026
Reviewed-by: Hudson CI
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>