symbols: Let MappingFile return a symbol with address
The MappingFile class now returns a symbol object with base address and
symbol name, instead of exposing its mapping and letting the provider
find the right entry. This will allow other symbol providers to
use this class as well, and also allow the provider to select the symbol
closest to the requested address if multiple files return a result.

Change-Id: Ic6a43d6b1daa0131ba6a1b680e0b0434bd4e4283
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/102631
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>